+++ CHANGELOG.rst
@@ -10,6 +10,10 @@ and this project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.
 Unreleased
 ----------
 
+**Fixed**
+
+* Resolve reconnection issues.
+
 0.4.0
 ----------